pyproject.toml
[tool.poetry]
name = "dw-blog-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.3"
pytest-django = "4.5.2"
python-dotenv = "^1.0.1"
black = "^24.3.0"
flake8 = "^7.0.0"
pytest-cov = "4.0.0"
factory-boy = "^3.3.0"
pytest-factoryboy = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
